requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
redis==5.0.8
pypdf==4.3.1
google-generativeai==0.8.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
pdfplumber==0.11.4